minLength/maxLength constraints.
Demonstrates a pattern of raising a ConstraintNotSupported exception in
the base class, which is overridden in subclasses that support the
constraint. In this example, only sequences like string, dict, and list
support max/minLength. Support is provided by mixing in the overriding
methods from a `constraints` module.

Most constraint code is in the `constraints` modules and mixed into the
type classes where needed.